Only try to insert if there are dependencies

+++ alembic/versions/2016111523_collapse_worker_dependency__540856fade99.py
@@ -21,20 +21,21 @@ def upgrade():
     op.add_column('worker', sa.Column('is_active', sa.Boolean(), nullable=False))
 
     # Convert table to new format
+    op.add_column('worker_dependency', sa.Column('dependencies', sa.LargeBinary(), nullable=False))
     conn = op.get_bind()
     results = conn.execute("SELECT * FROM worker_dependency").fetchall()
-    dependencies = {}
-    for row in results:
-        dependencies.setdefault((row.user_id, row.worker_id), []).append((row.dependency_uuid, row.dependency_path))
-    op.execute("DELETE FROM `worker_dependency`;")
-    op.execute(worker_dependency.insert().values([{
-        'user_id': user_id,
-        'worker_id': worker_id,
-        'dependencies': WorkerModel._serialize_dependencies(deps),
-    } for (user_id, worker_id), deps in dependencies.iteritems()]))
+    if results:
+        dependencies = {}
+        for row in results:
+            dependencies.setdefault((row.user_id, row.worker_id), []).append((row.dependency_uuid, row.dependency_path))
+        op.execute("DELETE FROM `worker_dependency`;")
+        op.execute(worker_dependency.insert().values([{
+            'user_id': user_id,
+            'worker_id': worker_id,
+            'dependencies': WorkerModel._serialize_dependencies(deps),
+        } for (user_id, worker_id), deps in dependencies.iteritems()]))
 
     # Use worker_dependency as a key-value store, with serialized dependencies in the MEDIUMBLOB
-    op.add_column('worker_dependency', sa.Column('dependencies', sa.LargeBinary(), nullable=False))
     op.drop_column('worker_dependency', 'dependency_uuid')
     op.drop_column('worker_dependency', 'dependency_path')
